Simplified registration of game actions and avoid duplicate code.
